Initial commit to move to ENTRYPOINT as root. WIP.
We need to deal with two, rather conflicting requirements #90 and

For k8s and other orchestration systems, there is no direct equivalent
to docker native volumes[1] but instead volumes are owned by root by
default. The GID at least can be influenced with `securityContext`[2],
but for lots of people it's easier to have an easier mechanism. One
solution is described in [3].

However the existing setup doesn't work if the container is executed
with a different `--user` or in Openshift where the `USER` in the
`Dockerfile` is overriden and the ENTRYPOINT/CMD is executed by a
random UID>=1024, member of 0 group.

This commit switches away from the esdocker CMD to an ENTRYPOINT.
The ENTRYPOINT (`/usr/local/bin/docker-entrypoint.sh`) combines Docker
+ Kubernetes + Openshift best practices.

The group ownership of `/usr/share/elasticsearch` is now 0, so in the
case of Openshift, or forcing `--user <other_uid>`, elasticsearch can
still run and write to bind mounted directories. The root group
doesn't grant any additional special permissions, so should somehow the
elasticsearch process in the container be compromised, it can't access/mutate
critical files from the host os, even if `/` is mounted in the container.

Optionally, a `MUTATE_MOUNTS` env var can be explicitly used to
override the UID:GID of the `data` and `logs` dirs as well, when the
ENTRYPOINT runs as the default root user (i.e. not in Openshift).

It's still possible to run another CMD, through the ENTRYPOINT, such
as `/bin/bash` to inspect things or directly `elasticsearch -Ex.y=z`[4].

Finally, CTRL-C is now handled more appropriately, when using the
image interactively.

[1]: https://success.docker.com/KBase/Different_Types_of_Volumes
[2]: https://github.com/kubernetes/kubernetes/issues/2630#issuecomment-323147775
[3]: https://github.com/elastic/elasticsearch-docker/issues/114#issuecomment-330248096
[4]: https://www.elastic.co/guide/en/elasticsearch/reference/current/docker.html#_d_override_the_image_8217_s_default_ulink_url_https_docs_docker_com_engine_reference_run_cmd_default_command_or_options_cmd_ulink